Better type system loading integration
