+++ packages/hash/frontend/src/lib/use-init-type-system.ts
@@ -1,18 +1,31 @@
 import init from "@blockprotocol/type-system-web";
-import { useEffect, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 
-export const useInitTypeSystem = () => {
+export const useAdvancedInitTypeSystem = () => {
   const [loadingTypeSystem, setLoadingTypeSystem] = useState(true);
+  const loadingPromise = useRef<Promise<void> | null>(null);
+
+  const loadTypeSystem = useCallback(() => {
+    if (loadingPromise.current) {
+      return loadingPromise.current;
+    }
+
+    loadingPromise.current = (async () => {
+      await init().then(() => {
+        setLoadingTypeSystem(false);
+      });
+    })();
+
+    return loadingPromise.current;
+  }, []);
 
   useEffect(() => {
     if (loadingTypeSystem) {
-      void (async () => {
-        await init().then(() => {
-          setLoadingTypeSystem(false);
-        });
-      })();
+      void loadTypeSystem();
     }
-  }, [loadingTypeSystem, setLoadingTypeSystem]);
+  }, [loadTypeSystem, loadingTypeSystem]);
 
-  return loadingTypeSystem;
+  return [loadingTypeSystem, loadTypeSystem] as const;
 };
+
+export const useInitTypeSystem = () => useAdvancedInitTypeSystem()[0];
